Issue 1106: [PATCH] Fixed potential usage of uninitialized bool variable in textord/tospace.cpp
https://code.google.com/p/tesseract-ocr/issues/detail?id=1106

Reported by ettl.martin78, Feb 11, 2014
Please review the attached patch. It fixes a potential usage of an uninitialized bool variable ('fuzzy_sp'). The fix simply initialized the variable by default to 'FALSE'. Before the fix, the value of 'fuzzy_sp' was not set, but used in the else-branch in line 1107:

          else {
            prev_blanks = blanks;
            prev_fuzzy_sp = fuzzy_sp;
            prev_fuzzy_non = fuzzy_non;
          }

Best regards and many thanks

Feb 11, 2014
#1 ettl.martin78
The updated patch fixes two more uninitialized variable usages in the same function.

Many thanks for reviewing.