Issue 1199: Remove the vestigal traces of custom memory allocators: 2. memry and emalloc removal.
2. memry and emalloc removal.

As above, but also removes emalloc* and replaces all uses of E{malloc,realloc,free} with the standard versions.

This is somewhat more controversial, because the E variants bothered to check the return value from the respective function, and would perform DoError if it determined it to have failed, which terminates in message to console and err_exit(). This is arguably nicer than crashing on null pointer value, but large parts of the code have never cared about malloc's potential NULL values, so I'd say it doesn't really help that parts of it now do.

Perhaps the real plan for moving forwards is to also replace malloc/free, and instead start using new/delete operators, and allow them to throw exceptions if the object construction fails due to memory allocation error. In this sense, the second patch may be considered as a stepping stone on the path towards such future version of Tesseract.

Conflicts:
	classify/flexfx.cpp