Add a command-line argument to set visible PDF image
Fix issue #210.

This adds an optional command-line argument to set the image which
will be used when generating a PDF image.

This addresses a niche case where the user wishes to use an optimized
image for OCR but maintain the visual appearance of the original image
when generating a PDF.

Signed-off-by: David Hendricks <david.hendricks@gmail.com>