Implement SIMD detection on macOS-gcc; complain if no SIMD detection
Verified that this compiles and runs on macOS El Capitan with homebrew gcc. I
don't believe clang compiles at the moment so not sure if it works for that.

Also added an #error if SIMD checking is missing. Platforms that have no
SIMD it seems to be should make that explicit.