Stage packages that provides the depending libraries
Snapcraft will soonly drop the feature to automatic staging the
depending libraries into the snap, specify it manually to avoid breakage
in the future.

This deals with the following warning message in the build:

```
Priming leptonica
Files from the build host were migrated into the snap to satisfy
dependencies that would otherwise not be met. This feature will be
removed in a future release. If these libraries are needed in the final
snap, ensure that the following are either satisfied by a stage-packages
entry or through a part:
usr/lib/x86_64-linux-gnu/libjbig.so.0
usr/lib/x86_64-linux-gnu/libjpeg.so.8
usr/lib/x86_64-linux-gnu/libopenjp2.so.7
usr/lib/x86_64-linux-gnu/libtiff.so.5
Priming tesseract
Files from the build host were migrated into the snap to satisfy
dependencies that would otherwise not be met. This feature will be
removed in a future release. If these libraries are needed in the final
snap, ensure that the following are either satisfied by a stage-packages
entry or through a part:
usr/lib/x86_64-linux-gnu/libgomp.so.1
```

Signed-off-by: 林博仁(Buo-ren, Lin) <Buo.Ren.Lin@gmail.com>