Add configure based detection for NEON runtime detection.
Previously, we assumed that if the compiler accepted
-mfpu=neon that we'd either be on ANDROID, or asm/hwcap.h
would be available to us.

There will be non-Android systems where asm/hwcap.h is not
available (indeed, the OS-FUZZ target seems to be one).

So, extend the checks to see whether asm/hwcap.h etc is
available. If it is not, then we don't attempt to use it,
and instead look at NEON_ALWAYS_AVAILABLE to decide whether
to hardwire support on/off.