Fix SIMD architecture detection logic.
Originally, this code would have complained if it was ever compiled
on a platform that didn't support it.

We changed this so that every file could be built on every platform
for simplicity of build files. Attempting to build (say) an SSE
file on a platform that didn't support SSE will just compile away
to nothing.

Unfortunately, while making this change, I didn't remove the
slightly strange state whereby it would be impossible to build
without SSE optimisations on a platform that supported them.

To fix this, I've removed the lines.