Fix linking Tesseract to project using CMake on Linux
In the past Tesseract library was wrongly named as "liblibtesseract" so it was needed to use "-llibtesseract" as argument for linker. The name was fixed in commit 52cac3a42ef2b823b9367b9ee2d1d84f8eda4ea0 but the ${Tesseract_LIBRARIES} variable still wrongly holds "libtesseract" and causes linker error when one is using it in target_link_libraries(...). This commit fixes that.