Fix linker flags for MSYS2 clang64 builds
MSYS2 clang64 uses the lld linker which does not support --as-needed.
The normal GNU ld uses that linker option with ELF targets but ignores
it for PE targets (.exe, .dll), so it can be removed.

Remove also the -Wl, which is only needed when linker options are
passed to the compiler but not when they are directly passed to the
linker.

Signed-off-by: Stefan Weil <sw@weilnetz.de>